# docs/_generated/studio_openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StudioTrainingWeightAttachRequest": {
        "description": "Request body for admin training weight-restore warm-start attach.",
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "expected_config_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Config Sha256"
          },
          "source_job_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Source Job Id",
            "type": "string"
          }
        },
        "required": [
          "source_job_id"
        ],
        "title": "StudioTrainingWeightAttachRequest",
        "type": "object"
      },
...
      "StudioTrainingWeightRestoreRequest": {
        "description": "Request body for admin training weight-restore materialization.",
        "properties": {
          "expected_config_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Config Sha256"
          },
          "source_job_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Source Job Id",
            "type": "string"
          }
        },
        "required": [
          "source_job_id"
        ],
        "title": "StudioTrainingWeightRestoreRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SC-NeuroCore Studio",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/studio/training/weight-restore": {
      "post": {
        "description": "Materialize and verify a training job's weights as confined evidence.\n\nBuilds the canonical restore plan from the source training job's stored\ncheckpoint metadata, fetches the integrity-checked weight and metadata\nartifacts, then runs the untrusted torch deserialization inside a bounded\nworker job. The job emits a path-free\n``studio.training.weight-restore.v1`` evidence artifact carrying only the\nverified digests and loaded-key totals; the in-memory tensor state\ndictionary never leaves the worker.",
        "operationId": "api_studio_training_weight_restore_api_studio_training_weight_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudioTrainingWeightRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Studio Training Weight Restore Api Studio Training Weight Restore Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Studio Training Weight Restore"
      }
    },
    "/api/studio/training/weight-restore/attach": {
      "post": {
        "description": "Warm-start a training job seeded with restored, verified weights.\n\nBuilds the canonical restore plan from the source job's checkpoint,\ndelivers the integrity-checked weights to a bounded process worker as\nconfined seed inputs, and starts a training job that loads them at the\nepoch-zero checkpoint boundary before training forward. A strict load of\nan incompatible architecture fails the job before training begins. The\nworker writes a path-free ``studio.training.weight-restore-attach.v1``\nevidence artifact; the deserialized tensors never reach the response.",
        "operationId": "api_studio_training_weight_restore_attach_api_studio_training_weight_restore_attach_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudioTrainingWeightAttachRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Studio Training Weight Restore Attach Api Studio Training Weight Restore Attach Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Studio Training Weight Restore Attach"
      }
    },
...
  }
}